Fix: SimpleCov errors out in 0.17.x because of missing option
Missed that in the last PR, because I was under the assumption that
SonarCloud was simply not reporting any coverage because there was no
code change. Figures that again `coverage/.resultset.json` wasn't
generated, because of using an non-existant SimpleCov configuration
property, that still existed in 0.19.x.

We can simply ignore that special subprocess handling, since our CLI
specs are actually running in-process.